refactor: move the meta-extractor implementation to its own file
